Project/requirements.txt
streamlit==1.36.0
tiktoken==0.7.0
langchain==0.2.10
langchain-community==0.2.9
pinecone-client==4.1.2
openai==1.34.0